Separate module and rename PubSub to MQ
