+++ .gitignore
@@ -1,0 +1,2 @@
+*.swp
+*__pycache__/
