+++ pubsubs/__init__.py
@@ -1,1 +1,3 @@
+from pubsubs.mq import MessageQueue
+
 from . import backends
